Added optional to types that can be null from the database.
When using functions it looks like doobie can't for sure determent
if the value can't be null or not. This makes it the api a bit dirty and
I prefer that we could find another workaround for his. I tried to wrap
the avg calls into a coalesce but than again it does not ensure that it
will not return null :/ The strange thing is that the analysis specs nor
 the integration specs had this problem.