fix: add Accept-Encoding gzip to sleeping pill requests
This is a regression from sleeping pill where they tried to add gzip
to the responses. So they (as described in the sparkjava documentation doh!)
add the header saying that the content is gzipped, regardless of if it's
done or not. The nice side effect of this is that clients that actually
read and actually use the headers will try and fail to unzip the content.